Version 0.3.0 - support per second templates if required.
Previously lumberjack was limited to per minute log files as the main
loop would only be executed every minute if there were no log lines to
be written.  The timeout is now calculated as a fraction of the second,
causing the read to timeout every second if there are no log lines to
write.  This allows the template to roll over every second if
necessary, allowing per second logfiles.